requirements.txt
discord.py==1.5.0
tortoise-orm==0.16.14
requests==2.24.0
aioredis==1.3.1
pytz==2020.1
pyseeyou==1.0.0
aiomysql==0.0.20
emoji==0.6.0
